requirements.txt
streamlit
pybaseball
pandas
google-genai
python-dotenv
plotly
